Fix wheel tagging, missing deps, and packaging gaps from review
Address code-review findings on the CI modernization PR, plus issues
found while verifying the built wheel:

- Wheels were mistagged py3-none-any despite containing compiled
  extensions: ext_modules were injected from a build_py subclass, after
  bdist_wheel had already decided the wheel was pure. Declare them
  eagerly in a new setup.py (metadata stays in pyproject.toml) and drop
  _build.py. The release workflow now import-tests a compiled module so
  a mistagged wheel fails the build.
- Declare scikit-image (hazma.theory) as a runtime dependency, and
  replace the importlib_resources backport (hazma.limits) and
  pkg_resources (hazma.cmb, hazma.pbh) with stdlib importlib.resources.
- Cap requires-python at < 3.13: numpy < 2 has no 3.13 support, so pip
  on newer Pythons attempted a doomed source build.
- Restrict package discovery to hazma*: namespace discovery with
  where = ['.', 'hazma'] shipped docs/, test/, notebooks/, and
  top-level duplicates of every subpackage in the wheel (17 MB -> 8.8 MB).
- Ship hazma/limits/data/*.dat, which had no package-data entry; the
  wheel previously broke 'import hazma.limits'.
- Pin build-time scipy to the runtime range: the extensions cimport
  scipy.special.cython_special, and building against scipy >= 1.14
  produced wheels whose C symbols don't exist in the scipy < 1.14
  runtime.
- CI now runs an import smoke test of the main public entry points from
  outside the repo; the test suite alone misses all of the above.

Verified on Python 3.12: wheel contains only hazma/ and dist-info with
correct cp312 platform tags, all public entry points import from the
installed wheel, and the test suite passes (57 passed, 10 skipped).
